pyproject.toml
[project]
name = "metro-sp-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.5",
    "numpy>=2.1.2",
    "opencv-python>=4.10.0.84",
    "weaviate-client>=4.0.0",
    "uvicorn>=0.32.0",
    "pydantic-settings>=2.6.1",
    "face-recognition>=1.3.0",
    "pillow>=11.0.0",
]


[tool.hatch.build.targets.wheel]
packages = ["src/main"]

[tool.uv.sources]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
